+++ lib/features/settings/ui/settings_screen.dart
@@ -4,6 +4,7 @@ import 'package:storii/app/init.dart';
 import 'package:storii/features/settings/ui/account_section.dart';
 import 'package:storii/features/settings/ui/app_cache_tile.dart';
 import 'package:storii/features/settings/ui/appearance/appearance_screen.dart';
+import 'package:storii/features/settings/ui/backup_settings_tile.dart';
 import 'package:storii/features/settings/ui/datetime_format_tile.dart';
 import 'package:storii/features/settings/ui/http_logs.dart';
 import 'package:storii/features/settings/ui/library/library_settings_screen.dart';
@@ -42,8 +43,8 @@ class SettingsScreen extends StatelessWidget {
           HttpLogsTile(),
           AppCacheTile(),
           TrustAllCertificatesTile(),
+          BackupSettingsTile(),
           AccountSection(),
-          // TODO: export/import settings
         ],
       ),
     );